fix:轨迹功能重做 add:1.增加表单保存到服务器功能     2.权限功能

+++ app/src/main/java/cn/cdjzxy/monitoringassistant/utils/PrintUtil.java
@@ -120,7 +120,7 @@ public class PrintUtil {
 
         //边框-监测项目
         sy = ey;
-        ey += (int) (normalHeight * 2);
+        ey += normalHeight * 2;
         tsc.addBox(minX, sy, maxX, ey, thickness);
 
         // 绘制-监测项目
